Add jobs/worker status route
