+++ conf/application.conf
@@ -76,7 +76,7 @@ features {
   autoBrushReadyDatasets = []
   taskReopenAllowedInSeconds = 30
   allowDeleteDatasets = true
-  jobsEnabled = false
+  jobsEnabled = true
   # For new users, the dashboard will show a banner which encourages the user to check out the following dataset.
   # If isDemoInstance == true, `/createExplorative/hybrid/true` is appended to the URL so that a new tracing is opened.
   # If isDemoInstance == false, `/view` is appended to the URL so that it's opened in view mode (since the user might not
